https://issues.apache.org/jira/browse/AMQ-3422 - bitarray serialization
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@1151857 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/util/BitArray.java
@@ -134,7 +134,7 @@ public class BitArray implements Serializable {
     public void writeToStream(DataOutput dataOut) throws IOException {
         dataOut.writeByte(length);
         if (length <= BYTE_SIZE) {
-            dataOut.writeByte((int)bits);
+            dataOut.writeByte((byte)bits);
         } else if (length <= SHORT_SIZE) {
             dataOut.writeShort((short)bits);
         } else if (length <= INT_SIZE) {
